+++ tests/ByteDev.Http.UnitTests/TestDummy.cs
@@ -1,0 +1,7 @@
+﻿namespace ByteDev.Http.UnitTests
+{
+    public class TestDummy
+    {
+        public string Name { get; set; }
+    }
+}